Moved home and about to pages folder.
